Spring MVC Beginner's Guide

+++ webstore/src/main/java/com/next/domain/Product.java
@@ -1,12 +1,10 @@
 package com.next.domain;
 
-import java.math.BigDecimal;
-
 public class Product {
 	
 	private String productId;
 	private String name;
-	private BigDecimal unitPrice;
+	private int unitPrice;
 	private String description;
 	private String manufacturer;
 	private String category;
@@ -19,7 +17,7 @@ public class Product {
 		super();
 	}
 	
-	public Product(String productId, String name, BigDecimal unitPrice) {
+	public Product(String productId, String name, int unitPrice) {
 		this.productId = productId;
 		this.name = name;
 		this.unitPrice = unitPrice;
@@ -41,11 +39,11 @@ public class Product {
 		this.name = name;
 	}
 
-	public BigDecimal getUnitPrice() {
+	public int getUnitPrice() {
 		return unitPrice;
 	}
 
-	public void setUnitPrice(BigDecimal unitPrice) {
+	public void setUnitPrice(int unitPrice) {
 		this.unitPrice = unitPrice;
 	}
